requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
pydantic
langchain-google-genai
